optimize RHFCountrySelect
update helperText in autocomplete form
